pyproject.toml
[tool.poetry]
name = "dynamicrender"
version = "0.1.0"
description = ""
authors = ["DMC <lzxder@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.6.0"
pillow = "^9.4.0"
qrcode = "^7.3.1"
httpx = "^0.23.2"
numpy = "^1.24.1"
dynamicadaptor = "0.1.6"


[[tool.poetry.source]]
name = "url"
url = "https://pypi.org/simple/"
default = false
secondary = false


[[tool.poetry.source]]
name = "secondary"
url = "https://pypi.mirrors.ustc.edu.cn/simple/"
default = false
secondary = true

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
